+++ src/Network/Ethereum/Ens.hs
@@ -24,6 +24,7 @@ import           Data.ByteArray                      (convert, zero)
 import           Data.ByteArray.Sized                (unsafeFromByteArrayAccess)
 import           Data.ByteString                     (ByteString)
 import           Data.ByteString.Char8               (split)
+import           Data.Monoid                         ((<>))
 import           Lens.Micro                          ((.~))
 
 import           Data.Solidity.Prim                  (Address, BytesN)